update step download, crop and annotate

+++ crop.py
@@ -5,7 +5,7 @@ import os
 import argparse
 
 
-def clip_text(captcha):
+def crop_text(captcha):
     text = captcha[0:28, 116:]
     gray = cv2.cvtColor(text, cv2.COLOR_BGR2GRAY)
 
@@ -19,7 +19,7 @@ def clip_text(captcha):
 
     return roi
 
-def clip_image(captcha):
+def crop_image(captcha):
     x_offset = 5
     y_offset = 41
     x_margin = 5
@@ -38,24 +38,38 @@ def clip_image(captcha):
     return images
 
 if __name__ == '__main__':
-    parser = argparse.ArgumentParser(description='parse and save text part of captcha')
-    parser.add_argument('-c', '--captcha', type=str, required=True, help='captcha directory to parse')
+    parser = argparse.ArgumentParser(description='crop captcha to text and image part')
+    parser.add_argument('-d', '--dir', type=str, required=True, help='captcha directory')
     parser.add_argument('-t', '--text', type=str, required=True, help='target directory to save text part')
     parser.add_argument('-i', '--image', type=str, required=True, help='target directory to save image part')
     args = vars(parser.parse_args())
 
-    image_paths = list(paths.list_images(args["captcha"]))
+    os.makedirs(args['text'], exist_ok=True)
+    os.makedirs(args['image'], exist_ok=True)
+
+    image_paths = list(paths.list_images(args["dir"]))
+    
     for captcha_path in image_paths:
-        basename = os.path.basename(captcha_path)
-        (filename, ext) = os.path.splitext(basename)
+        captcha_name = os.path.basename(captcha_path)
+        (serial, ext) = os.path.splitext(captcha_name)
 
         captcha = cv2.imread(captcha_path)
-        text = clip_text(captcha)
-        text_path = os.path.join(args["text"], basename)
-        cv2.imwrite(text_path, text)
+        text = crop_text(captcha)
+        text_path = os.path.join(args["text"], captcha_name)
 
-        images = clip_image(captcha)
+        if os.path.exists(text_path):
+            print("already saved captcha %s text part." % captcha_name)
+        else:
+            cv2.imwrite(text_path, text)
+            print("save captcha %s text part to %s." % (captcha_name, text_path))
+
+        images = crop_image(captcha)
         for i, img in enumerate(images):
-            img_name = '{}_{}{}'.format(filename, str(i), ext)
+            img_name = '{}_{}{}'.format(serial, str(i), ext)
             img_path = os.path.join(args["image"], img_name)
-            cv2.imwrite(img_path, img)
+
+            if os.path.exists(img_path):
+                print("already saved captcha %s image part %d." % (captcha_name, i))
+            else:
+                cv2.imwrite(img_path, img)
+                print("save captcha %s image part %d to %s." % (captcha_name, i, img_path))
